requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
python-dotenv==1.0.1
playwright==1.42.0
jupyter==1.0.0
matplotlib==3.7.1
seaborn==0.12.2
Pillow==10.2.0
black==24.3.0
isort==5.13.2
tqdm==4.66.2
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.2.1+cu121
torchvision==0.17.1+cu121
timm==0.9.12
catboost
xgboost
lightgbm
joblib
shap
notebook 